% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_authors.R
\name{zb_authors}
\alias{zb_authors}
\alias{zb_authors_}
\title{Search for authors}
\usage{
zb_authors(id = NULL, query = NULL, parse = TRUE, ...)

zb_authors_(id = NULL, query = NULL, ...)
}
\arguments{
\item{id}{(integer/numeric) An author identifier}

\item{query}{(character) Query terms}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). If \code{TRUE}, we also give back a
tibble for easier consumption. Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
Search for authors
}
\examples{
\dontrun{
zb_authors(id = "8C466CBE-3F7D-4DC9-8CBD-26DD3F57E212")
zb_authors(query = "Schmutz")
zb_authors(query = "Pyle")
}
}
