% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_id.R, R/zb_matching.R
\name{zb_id}
\alias{zb_id}
\alias{zb_id_}
\alias{zb_matching_}
\title{Get data by identifier}
\usage{
zb_id(id, parse = TRUE, ...)

zb_id_(id, ...)

zb_matching_(id, ...)
}
\arguments{
\item{id}{(integer/numeric) Any ZooBank identifier, for taxon, author
or publication. required}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). If \code{TRUE}, we also give back a
tibble for easier consumption. Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
Get data by identifier
}
\examples{
\dontrun{
zb_id(id = "6EA8BB2A-A57B-47C1-953E-042D8CD8E0E2")
}
}
