% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_matching.R
\name{zb_matching}
\alias{zb_matching}
\title{Matching taxon name service}
\usage{
zb_matching(id, parse = TRUE, ...)
}
\arguments{
\item{id}{(integer/numeric) Any ZooBank identifier, for taxon, author
or publication. required}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). If \code{TRUE}, we also give back a
tibble for easier consumption. Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
Matching taxon name service
}
\examples{
\dontrun{
zb_matching(id = "FFF7160A-372D-40E9-9611-23AF5D9EAC4C")
}
}
