% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_name_usages.R
\name{zb_name_usages}
\alias{zb_name_usages}
\alias{zb_name_usages_}
\title{Name usages}
\usage{
zb_name_usages(name = NULL, id = NULL, query = NULL, parse = TRUE,
  ...)

zb_name_usages_(name = NULL, id = NULL, query = NULL, ...)
}
\arguments{
\item{name}{(character) A taxonomic name}

\item{id}{(integer/numeric) A taxonomic identifier}

\item{query}{(character) A taxonomic name to query}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). If \code{TRUE}, we also give back a
tibble for easier consumption. Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
Name usages
}
\examples{
\dontrun{
zb_name_usages(name = "Pseudanthias carlsoni")
zb_name_usages(id = "6EA8BB2A-A57B-47C1-953E-042D8CD8E0E2")
zb_name_usages(query = "pyle")
}
}
