% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zb_publications.R
\name{zb_publications}
\alias{zb_publications}
\alias{zb_publications_}
\title{Publications}
\usage{
zb_publications(id = NULL, query = NULL, parse = TRUE, ...)

zb_publications_(id, query, ...)
}
\arguments{
\item{id}{(integer/numeric) A publication identifier}

\item{query}{(character) Query terms}

\item{parse}{(logical) Whether to parse to list (\code{FALSE}) or
data.frame (\code{TRUE}). If \code{TRUE}, we also give back a
tibble for easier consumption. Default: \code{TRUE}}

\item{...}{Curl options passed to \code{\link[crul]{HttpClient}}}
}
\description{
Publications
}
\examples{
\dontrun{
zb_publications(id = "427D7953-E8FC-41E8-BEA7-8AE644E6DE77")
zb_publications(query = "pyle")
}
}
