% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zbank-package.R
\docType{package}
\name{zbank-package}
\alias{zbank-package}
\alias{zbank}
\title{zbank - ZooBank Client}
\description{
zbank - ZooBank Client
}
\section{ZooBank API Docs}{

See \url{http://zoobank.org/Api}
}

\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
